Removed conflicts markers for hillas.py
For all the changes commited in hillas.py please see the changelog at the top of the code. This code is currently using hillas_parameters_1 by default. A parallel code development and implementation of higher order image parameters in hillas_parameters_2 will be done after removal of conflicts between the two versions (the discrepancy in length and width parameters between the two versions is already fixed in this code).