New mc.py
The new mc.py is extremely simple, as the only operation unique to MC data is the conversion from r0 to dl0, as the data is obtained as r0 from hessio files. In the future this functionality will be called at IO, and dl0 will correctly contain the photoelectron samples that we will obtain from real data files.
mc_r0_to_dl0_calibration simply subtracts the MC pedestal, and applies the MC dc_to_pe, resulting in a numpy array of photoelectron samples.